In progress

+++ openrefine/views.py
@@ -20,15 +20,20 @@ def openrefine(request):
     if request.method == "POST":
         form = OpenRefineForm(request.POST)
         print("Ready to Process ", ready4or)
-        try:
-            response = openRefineSteps(ready4or)
-        except BaseException as error:
-            response = 'FAIL: An error occurred: {}'.format(error)
 
-        nextReady = False if 'FAIL' in response else True
+        if ready4or == '':
+            nextReady = False
+            return render(request, 'openrefine/openrefine.html', {'nextReady': nextReady })
+        else:
+            try:
+                response = openRefineSteps(ready4or)
+            except BaseException as error:
+                response = 'FAIL: An error occurred: {}'.format(error)
 
-        [f.unlink() for f in Path(settings.MEDIA_ROOT + '/').glob("*") if f.is_file()]
-        return render(request, 'openrefine/openrefine.html', {'nextReady': nextReady, 'response': response})
+            nextReady = False if 'FAIL' in response else True
+
+            [f.unlink() for f in Path(settings.MEDIA_ROOT + '/').glob("*") if f.is_file()]
+            return render(request, 'openrefine/openrefine.html', {'nextReady': nextReady, 'response': response})
     else:
         form = OpenRefineForm
